prepare adding polyglot concept
