+++ test/http-redirection/http-polyglot.redirection.test.js
@@ -2,9 +2,9 @@ import { assert } from "@jsenv/assert"
 import { jsenvPrivateKey, jsenvCertificate, fetchUrl } from "@jsenv/server"
 import { headersToObject } from "@jsenv/server/src/headersToObject.js"
 import { listen } from "@jsenv/server/src/internal/listen.js"
-import { createServer } from "./http-polyglot.js"
+import { createPolyglotServer } from "@jsenv/server/src/internal/server-polyglot.js"
 
-const server = createServer({
+const server = createPolyglotServer({
   privateKey: jsenvPrivateKey,
   certificate: jsenvCertificate,
   requestHandler: (nodeRequest, nodeResponse) => {
